[feat] 사용자 댓글 삭제 기능 추가 - 수정

+++ core-service/src/main/java/com/microsoftwo/clother/comment/command/application/dto/CommentDTO.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class CommentDTO {
-    private Long id;
-    private Long userId;
-    private Long postId;
-    private Long boardId;
-    private Long parentId;
+    private int id;
+    private int userId;
+    private int postId;
+    private int boardId;
+    private int parentId;
     private String content;
     private LocalDateTime createdAt;
     private Boolean isDeleted = false;
