+++ core-service/src/main/java/com/microsoftwo/clother/comment/command/domain/aggregate/Comment.java
@@ -16,18 +16,18 @@ public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // 댓글 ID
+    private int id; // 댓글 ID
 
     @Column(nullable = false)
-    private Long userId; // 회원 ID
+    private int userId; // 회원 ID
 
 
-    private Long postId; // 포스트 ID (NULL 가능)
+    private int postId; // 포스트 ID (NULL 가능)
 
 
-    private Long boardId; // 게시판 ID (NULL 가능)
+    private int boardId; // 게시판 ID (NULL 가능)
 
-    private Long parentId; // 부모 댓글 ID (NULL 가능)
+    private int parentId; // 부모 댓글 ID (NULL 가능)
 
     @Column(nullable = false, length = 255)
     private String content; // 댓글 내용
